django_hands_on/requirements.txt
django==4.1.5
djangorestframework==3.14.0
requests==2.82.2